+++ react/OrderBy.js
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types'
 import React, { useMemo } from 'react'
-import { injectIntl, intlShape } from 'react-intl'
+import { useIntl } from 'react-intl'
 
 import SelectionListOrderBy from './components/SelectionListOrderBy'
 
@@ -42,16 +42,17 @@ export const SORT_OPTIONS = [
 const OrderBy = ({
   orderBy,
   message,
-  intl,
   hiddenOptions = [],
   showOrderTitle = true,
 }) => {
+  const intl = useIntl()
+
   const sortingOptions = useMemo(() => {
     return SORT_OPTIONS.filter(
-      option => !hiddenOptions.includes(option.value)
+      (option) => !hiddenOptions.includes(option.value)
     ).map(({ value, label }) => {
       return {
-        value: value,
+        value,
         label: intl.formatMessage({ id: label }),
       }
     })
@@ -70,14 +71,12 @@ const OrderBy = ({
 OrderBy.propTypes = {
   /** Which sorting option is selected. */
   orderBy: PropTypes.string,
-  /** Intl instance. */
-  intl: intlShape,
   /** Options to be hidden. (e.g. `["OrderByNameASC", "OrderByNameDESC"]`) */
   hiddenOptions: PropTypes.arrayOf(PropTypes.string),
   /** Message to be displayed */
   message: PropTypes.string,
   /** Show or hide order title */
-  showOrderTitle: PropTypes.boolean,
+  showOrderTitle: PropTypes.bool,
 }
 
-export default injectIntl(OrderBy)
+export default OrderBy