+++ react/components/FilterNavigator/legacy/AccordionFilterContainer.js
@@ -1,22 +1,21 @@
 import React, { useState } from 'react'
 import PropTypes from 'prop-types'
-import { injectIntl, intlShape } from 'react-intl'
+import { useIntl } from 'react-intl'
 import classNames from 'classnames'
 import { IconCaret } from 'vtex.store-icons'
 
 import AccordionFilterItem from './AccordionFilterItem'
-
 import searchResult from './searchResult.css'
 
 const AccordionFilterContainer = ({
   filters,
-  intl,
   onFilterCheck,
   isOptionSelected,
 }) => {
+  const intl = useIntl()
   const [openItem, setOpenItem] = useState(null)
 
-  const handleOpen = id => e => {
+  const handleOpen = (id) => (e) => {
     e.preventDefault()
 
     if (openItem === id) {
@@ -26,13 +25,13 @@ const AccordionFilterContainer = ({
     }
   }
 
-  const handleKeyDown = e => {
+  const handleKeyDown = (e) => {
     if (e.key === ' ') {
       setOpenItem(null)
     }
   }
 
-  const nonEmptyFilters = filters.filter(spec => spec.facets.length > 0)
+  const nonEmptyFilters = filters.filter((spec) => spec.facets.length > 0)
 
   return (
     <div className={`${searchResult.accordionFilter} h-100`}>
@@ -67,7 +66,7 @@ const AccordionFilterContainer = ({
         )}
       </div>
 
-      {nonEmptyFilters.map(filter => {
+      {nonEmptyFilters.map((filter) => {
         const { title, facets } = filter
         const isOpen = openItem === filter.title
 
@@ -91,8 +90,6 @@ const AccordionFilterContainer = ({
 AccordionFilterContainer.propTypes = {
   /** Current available filters */
   filters: PropTypes.arrayOf(PropTypes.object),
-  /** Intl instance */
-  intl: intlShape,
   /** Filters mapped for checkbox */
   filtersChecks: PropTypes.object,
   /** Checkbox hit callback function */
@@ -102,4 +99,4 @@ AccordionFilterContainer.propTypes = {
   isOptionSelected: PropTypes.func.isRequired,
 }
 
-export default injectIntl(AccordionFilterContainer)
+export default AccordionFilterContainer