+++ react/components/FilterNavigator/legacy/PriceRange.js
@@ -1,7 +1,7 @@
 import React, { useRef } from 'react'
 import PropTypes from 'prop-types'
 import { useRuntime } from 'vtex.render-runtime'
-import { injectIntl, intlShape } from 'react-intl'
+import { useIntl } from 'react-intl'
 import { Slider } from 'vtex.styleguide'
 
 import { facetOptionShape } from '../../../constants/propTypes'
@@ -14,11 +14,13 @@ import FilterOptionTemplate from './FilterOptionTemplate'
 const DEBOUNCE_TIME = 500 // ms
 
 /** Price range slider component */
-const PriceRange = ({ title, facets, intl, priceRange }) => {
+const PriceRange = ({ title, facets, priceRange }) => {
   const {
     navigate,
     culture: { currency },
   } = useRuntime()
+
+  const intl = useIntl()
 
   const navigateTimeoutId = useRef()
 
@@ -61,15 +63,15 @@ const PriceRange = ({ title, facets, intl, priceRange }) => {
   let maxValue = Number.MIN_VALUE
 
   availableOptions.forEach(({ slug }) => {
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const [_, minSlug, maxSlug] = slug.match(slugRegex)
+    const [, minSlug, maxSlug] = slug.match(slugRegex)
 
-    const min = parseInt(minSlug)
-    const max = parseInt(maxSlug)
+    const min = parseInt(minSlug, 10)
+    const max = parseInt(maxSlug, 10)
 
     if (min < minValue) {
       minValue = min
     }
+
     if (max > maxValue) {
       maxValue = max
     }
@@ -79,11 +81,10 @@ const PriceRange = ({ title, facets, intl, priceRange }) => {
   const currentValuesRegex = /^(.*) TO (.*)$/
 
   if (priceRange && currentValuesRegex.test(priceRange)) {
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const [_, currentMin, currentMax] = priceRange.match(currentValuesRegex)
+    const [, currentMin, currentMax] = priceRange.match(currentValuesRegex)
 
-    defaultValues[0] = parseInt(currentMin)
-    defaultValues[1] = parseInt(currentMax)
+    defaultValues[0] = parseInt(currentMin, 10)
+    defaultValues[1] = parseInt(currentMax, 10)
   }
 
   return (
@@ -96,7 +97,7 @@ const PriceRange = ({ title, facets, intl, priceRange }) => {
         max={maxValue}
         onChange={handleChange}
         defaultValues={defaultValues}
-        formatValue={value => intl.formatNumber(value, currencyOptions)}
+        formatValue={(value) => intl.formatNumber(value, currencyOptions)}
         range
       />
     </FilterOptionTemplate>
@@ -108,10 +109,8 @@ PriceRange.propTypes = {
   title: PropTypes.string.isRequired,
   /** Available price ranges */
   facets: PropTypes.arrayOf(facetOptionShape).isRequired,
-  /** Intl instance */
-  intl: intlShape.isRequired,
-  /** Current price range filter query parameter*/
+  /** Current price range filter query parameter */
   priceRange: PropTypes.string,
 }
 
-export default injectIntl(PriceRange)
+export default PriceRange