+++ react/components/FilterNavigator/legacy/SideBar.js
@@ -1,7 +1,9 @@
+/* eslint-disable jsx-a11y/no-static-element-interactions */
+/* eslint-disable jsx-a11y/click-events-have-key-events */
 import React, { Component } from 'react'
 import ReactDOM from 'react-dom'
 import PropTypes from 'prop-types'
-import { injectIntl, intlShape } from 'react-intl'
+import { injectIntl } from 'react-intl'
 import OutsideClickHandler from 'react-outside-click-handler'
 import Animation from 'vtex.store-components/Animation'
 import classNames from 'classnames'
@@ -34,15 +36,25 @@ class Sidebar extends Component {
 
   render() {
     const { isOpen, onOutsideClick } = this.props
+
     if (typeof document === 'undefined') {
       return null
     }
-    const scrimClasses = classNames('fixed dim bg-base--inverted top-0 z-9999 w-100 vh-100 o-40 left-0', {
-      dn: !isOpen,
-    })
+
+    const scrimClasses = classNames(
+      'fixed dim bg-base--inverted top-0 z-9999 w-100 vh-100 o-40 left-0',
+      {
+        dn: !isOpen,
+      }
+    )
+
     return ReactDOM.createPortal(
       <OutsideClickHandler onOutsideClick={onOutsideClick}>
-        <div style={{ willChange: 'opacity' }} className={scrimClasses} onClick={onOutsideClick} />
+        <div
+          style={{ willChange: 'opacity' }}
+          className={scrimClasses}
+          onClick={onOutsideClick}
+        />
         <Animation
           className={`${searchResult.sidebar} w-80 w-auto-ns h-100 fixed top-0 right-0 z-9999 bg-base shadow-2 flex flex-column`}
           isActive={isOpen}
@@ -58,7 +70,7 @@ class Sidebar extends Component {
 
 Sidebar.propTypes = {
   /* Internationalization */
-  intl: intlShape.isRequired,
+  intl: PropTypes.any,
   /* Set the sideBar visibility */
   isOpen: PropTypes.bool,
   /* Sidebar content */