+++ react/components/PriceRange.js
@@ -1,7 +1,7 @@
 import React, { useRef } from 'react'
 import PropTypes from 'prop-types'
 import { useRuntime } from 'vtex.render-runtime'
-import { injectIntl, intlShape } from 'react-intl'
+import { useIntl } from 'react-intl'
 import { Slider } from 'vtex.styleguide'
 import { formatCurrency } from 'vtex.format-currency'
 
@@ -13,8 +13,9 @@ import useSearchState from '../hooks/useSearchState'
 const DEBOUNCE_TIME = 500 // ms
 
 /** Price range slider component */
-const PriceRange = ({ title, facets, intl, priceRange }) => {
+const PriceRange = ({ title, facets, priceRange }) => {
   const { culture, setQuery } = useRuntime()
+  const intl = useIntl()
 
   const navigateTimeoutId = useRef()
 
@@ -47,15 +48,15 @@ const PriceRange = ({ title, facets, intl, priceRange }) => {
   let maxValue = Number.MIN_VALUE
 
   availableOptions.forEach(({ slug }) => {
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const [_, minSlug, maxSlug] = slug.match(slugRegex)
+    const [, minSlug, maxSlug] = slug.match(slugRegex)
 
-    const min = parseInt(minSlug)
-    const max = parseInt(Math.ceil(maxSlug))
+    const min = parseInt(minSlug, 10)
+    const max = parseInt(Math.ceil(maxSlug), 10)
 
     if (min < minValue) {
       minValue = min
     }
+
     if (max > maxValue) {
       maxValue = max
     }
@@ -65,11 +66,10 @@ const PriceRange = ({ title, facets, intl, priceRange }) => {
   const currentValuesRegex = /^(.*) TO (.*)$/
 
   if (priceRange && currentValuesRegex.test(priceRange)) {
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const [_, currentMin, currentMax] = priceRange.match(currentValuesRegex)
+    const [, currentMin, currentMax] = priceRange.match(currentValuesRegex)
 
-    defaultValues[0] = parseInt(currentMin)
-    defaultValues[1] = parseInt(currentMax)
+    defaultValues[0] = parseInt(currentMin, 10)
+    defaultValues[1] = parseInt(currentMax, 10)
   }
 
   return (
@@ -83,7 +83,7 @@ const PriceRange = ({ title, facets, intl, priceRange }) => {
         max={maxValue}
         onChange={handleChange}
         defaultValues={defaultValues}
-        formatValue={value => formatCurrency({ intl, culture, value })}
+        formatValue={(value) => formatCurrency({ intl, culture, value })}
         range
       />
     </FilterOptionTemplate>
@@ -95,10 +95,8 @@ PriceRange.propTypes = {
   title: PropTypes.string.isRequired,
   /** Available price ranges */
   facets: PropTypes.arrayOf(facetOptionShape).isRequired,
-  /** Intl instance */
-  intl: intlShape.isRequired,
-  /** Current price range filter query parameter*/
+  /** Current price range filter query parameter */
   priceRange: PropTypes.string,
 }
 
-export default injectIntl(PriceRange)
+export default PriceRange