+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- New props: `customSummaryInterval` and `CustomSummary` to `gallery` block.
+
+### Changed
+- Migrate legacy Gallery components to TypeScript.
+- Update usage of react-intl.
 
 ## [3.91.0] - 2021-01-11